# open-api/openapi-specs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/category": {
      "get": {
        "operationId": "getCategoryById",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCategoryDto"
                }
              }
            },
            "description": ""
          }
        }
      },
      "post": {
        "operationId": "createCategory",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCategoryDto"
                }
              }
            },
            "description": ""
          }
        }
      },
      "put": {
        "operationId": "updateCategory",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/statistic/merge/{range}": {
      "post": {
        "operationId": "merge",
        "parameters": [
          {
            "name": "range",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        }
      }
    },
    "/statistic/top-campaigns": {
      "get": {
        "operationId": "getTopCampaigns",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        }
      }
    },
    "/statistic/top-campaigns-category": {
      "get": {
        "operationId": "getTopCampaignsCategory",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        }
      }
    },
    "/statistic/{type}": {
      "post": {
        "operationId": "trackStatistic",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackStatisticDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        }
      }
    },
...
  },
  "info": {
    "title": "Aide",
    "description": "Aide API",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "aide_access_token"
      }
    },
    "schemas": {
...
      "CreateCategoryDto": {
        "properties": {
          "color": {
            "maxLength": 50,
            "type": "string"
          },
          "handle": {
            "maxLength": 200,
            "type": "string"
          },
          "mediaUrl": {
            "type": "string"
          },
          "subtitle": {
            "maxLength": 200,
            "type": "string"
          },
          "title": {
            "maxLength": 100,
            "type": "string"
          }
        },
        "required": [
          "color",
          "handle",
          "subtitle",
          "title"
        ],
        "type": "object"
      },
...
      "GetCategoryDto": {
        "properties": {
          "_id": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "handle": {
            "type": "string"
          },
          "mediaUrl": {
            "type": "string"
          },
          "order": {
            "type": "number"
          },
          "subtitle": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "_id",
          "color",
          "handle",
          "order",
          "subtitle",
          "title"
        ],
        "type": "object"
      },
...
      "TrackMetadataDto": {
        "properties": {
          "campaignId": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "TrackStatisticDto": {
        "properties": {
          "categoryClick": {
            "type": "number"
          },
          "click": {
            "type": "number"
          },
          "impression": {
            "type": "number"
          },
          "metadata": {
            "$ref": "#/components/schemas/TrackMetadataDto"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "metadata"
        ],
        "type": "object"
      },
...
      "UpdateCategoryDto": {
        "properties": {
          "_id": {
            "type": "string"
          },
          "color": {
            "maxLength": 50,
            "type": "string"
          },
          "handle": {
            "maxLength": 200,
            "type": "string"
          },
          "mediaUrl": {
            "type": "string"
          },
          "subtitle": {
            "maxLength": 200,
            "type": "string"
          },
          "title": {
            "maxLength": 100,
            "type": "string"
          }
        },
        "required": [
          "_id"
        ],
        "type": "object"
      },
...
    }
  }
}